originally i did a much more comprehensive list but I accidently deleted it unfortunatly.
--------------------

mild code rework
additon of stack_memory() - not working due to http error 500
error prevention to fix common errors related to missing or
incorrect api key or lack of message input

--------------------

need to fix memory stacking still. trying to stack memorys
from the api is throwing a HTTP error <500> (Internal Server Error)

want to look into also autocreation of .env file if it is nonexistant.
to make end user usage and setup easier, but not a priority.

want to clean code up a bit more. so far im relativly satisfied, especially
with my abuse of comments (mainly because they make it easier
for anyone, including those less familiar with python, to read and
understand the code)